Deep_learning_model_minimal-entropy-correlation-alignment/requirements.txt
numpy
scipy
pillow
tensorflow
scikit-image
scikit-learn
matplotlib



